# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WordMaster API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vocabularies/import": {
      "post": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Import Vocabulary",
        "operationId": "import_vocabulary_api_v1_vocabularies_import_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_vocabulary_api_v1_vocabularies_import_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vocabularies": {
      "get": {
        "tags": [
          "vocabularies"
        ],
        "summary": "List Vocabularies",
        "operationId": "list_vocabularies_api_v1_vocabularies_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VocabularyRead"
                  },
                  "title": "Response List Vocabularies Api V1 Vocabularies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vocabularies/{vocabulary_id}": {
      "get": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Get Vocabulary",
        "operationId": "get_vocabulary_api_v1_vocabularies__vocabulary_id__get",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Update Vocabulary",
        "operationId": "update_vocabulary_api_v1_vocabularies__vocabulary_id__patch",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Delete Vocabulary",
        "operationId": "delete_vocabulary_api_v1_vocabularies__vocabulary_id__delete",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Confirm"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vocabularies/{vocabulary_id}/export": {
      "get": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Export Vocabulary",
        "operationId": "export_vocabulary_api_v1_vocabularies__vocabulary_id__export_get",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vocabularies/{vocabulary_id}/active": {
      "patch": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Set Vocabulary Active",
        "operationId": "set_vocabulary_active_api_v1_vocabularies__vocabulary_id__active_patch",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyActiveUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vocabularies/{vocabulary_id}/priority": {
      "patch": {
        "tags": [
          "vocabularies"
        ],
        "summary": "Set Vocabulary Priority",
        "operationId": "set_vocabulary_priority_api_v1_vocabularies__vocabulary_id__priority_patch",
        "parameters": [
          {
            "name": "vocabulary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vocabulary Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyPriorityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_import_vocabulary_api_v1_vocabularies_import_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "name",
          "file"
        ],
        "title": "Body_import_vocabulary_api_v1_vocabularies_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportResponse": {
        "properties": {
          "vocabulary": {
            "$ref": "#/components/schemas/VocabularyRead"
          },
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          },
          "row_errors": {
            "items": {
              "$ref": "#/components/schemas/RowErrorRead"
            },
            "type": "array",
            "title": "Row Errors"
          }
        },
        "type": "object",
        "required": [
          "vocabulary",
          "imported_count",
          "row_errors"
        ],
        "title": "ImportResponse"
      },
...
      "RowErrorRead": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "code",
          "message"
        ],
        "title": "RowErrorRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VocabularyActiveUpdate": {
        "properties": {
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "active"
        ],
        "title": "VocabularyActiveUpdate"
      },
      "VocabularyPriorityUpdate": {
        "properties": {
          "priority": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "priority"
        ],
        "title": "VocabularyPriorityUpdate"
      },
      "VocabularyRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "total_words": {
            "type": "integer",
            "title": "Total Words"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "filename",
          "imported_at",
          "total_words",
          "active",
          "priority"
        ],
        "title": "VocabularyRead"
      },
      "VocabularyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "VocabularyUpdate"
      },
...
    }
  }
}